% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mean_function.R
\name{getMean}
\alias{getMean}
\title{A mean}
\usage{
getMean(x)
}
\arguments{
\item{x}{vector or matrix.}
}
\description{
This function allows you to get the mean of a vector/matrix
}
\examples{
getMean(c(1,2,3))
}

